Move query hgnc to view hgnc in commands
