+++ scout/commands/view/base.py
@@ -15,6 +15,7 @@ from .intervals import intervals
 from .collections import collections
 from .transcripts import transcripts
 from .case import cases
+from .hgnc import hgnc
 
 
 LOG = logging.getLogger(__name__)
